added package logic to app
